changing rating field from integer to float to facilitate a better rating

+++ mywatchlist/models.py
@@ -3,6 +3,6 @@ from django.db import models
 class myWatchList(models.Model):
     title = models.CharField(max_length=255)
     watched = models.CharField(max_length=3)
-    rating = models.IntegerField()
+    rating = models.FloatField()
     release_date = models.CharField(max_length=50)
     review = models.TextField()
